Extract WindowOp state machine into listeners::commands
Closes #37. Wave 3 extraction.

`listeners.rs` mixed two concerns: GTK-glue wiring (keyboard, focus
detector, file watcher, signal poller) and a pure state-machine for
window-command dispatch (`WindowOp` + `resolve_window_op` +
`handle_window_command`). The state machine has 7 unit tests covering
every WindowCommand × visible × resident combination — zero GTK deps,
fully testable. Pulled it into a sibling submodule so the file boundary
matches the conceptual one.

Layout after:

- `src/listeners.rs` (296 lines, was 437) — wiring only
- `src/listeners/commands.rs` (158 lines, new) — `WindowOp`,
  `resolve_window_op`, `handle_window_command`, `clear_drawer_state`,
  and the 7-test module that follows the code

`commands` is a child module of `listeners` so it can reach
`super::complete_show` and `super::quit_or_hide` without any visibility
changes (child modules see private items in their ancestors).

CLAUDE.md "What lives where" updated.

Zero behavior change. `make lint` clean. All 82 tests pass (including
the 7 moved ones).
