+++ CLAUDE.md
@@ -63,7 +63,9 @@ src/
 ├── config.rs            # clap CLI with CloseButton enum
 ├── state.rs             # DrawerState + AppRegistry sub-struct
 ├── desktop_loader.rs    # Scans .desktop files, multi-category assignment
-├── listeners.rs         # Focus detector, file watcher, signal receiver
+├── listeners.rs         # Focus detector, file watcher, signal receiver wiring
+├── listeners/
+│   └── commands.rs      # Pure WindowOp state machine + tests (zero GTK deps)
 ├── watcher.rs           # inotify watcher thread (.desktop + pin-file)
 └── ui/
     ├── navigation.rs     # install_grid_nav — capture-phase arrow-key
